Split GameActor into Player and Enemies
