+++ Project2Data/Project2.Models/Actor/ActorEnemy.cs
@@ -1,0 +1,16 @@
+using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
+
+namespace Project2.Models.Actor {
+    public class ActorEnemy : GameActor {
+        //  Default Constructor
+        public ActorEnemy() : base() {
+
+        }
+
+        //  Copy Constructor
+        public ActorEnemy(ActorEnemy pEnemy) : base(pEnemy) {
+            
+        }
+    }
+}
